Move image logic to parent component for event landing page's usage
