+++ app/(event)/event/[userId]/[eventId]/page.tsx
@@ -7,7 +7,13 @@ import { useEffect, useState } from 'react';
 import { twMerge } from 'tailwind-merge';
 import 'react-quill/dist/quill.snow.css';
 import Button from '@/components/Button';
-
+import {
+  ref,
+  uploadBytes,
+  getDownloadURL,
+  listAll,
+  getStorage
+} from 'firebase/storage';
 interface PageProps {
   userId: string;
   eventId: string;
@@ -21,6 +27,23 @@ export default function Template({ params }: { params: PageProps }) {
     useState<boolean>(true);
   //  previewMode will be set to false when users copy the iframe
   const [previewMode, setPreviewMode] = useState<boolean>(true);
+  // image URL to use here or pass down as prop to child components
+  const [imageUrls, setImageUrls] = useState<string[]>([]);
+
+  // storage and buckets as image reference
+  const storage = getStorage();
+  const bucket = ref(storage, `users/${params.userId}/${params.eventId}/1`);
+
+  // Retrieving Image logic
+  useEffect(() => {
+    listAll(bucket).then((response) => {
+      response.items.forEach((item) => {
+        getDownloadURL(item).then((url) => {
+          setImageUrls((prev) => [url]);
+        });
+      });
+    });
+  }, []);
 
   useEffect(() => {
     const fetchTemplate = async () => {
@@ -57,13 +80,18 @@ export default function Template({ params }: { params: PageProps }) {
             backgroundColor: data['backgroundColor']
           }}
         >
-          <Button
-            text="To Registration Form"
-            className="absolute right-4 top-4 rounded bg-blue-500 px-4 py-2 text-white hover:bg-blue-600"
-            onClick={() => {
-              setShowEventLandingPage(false);
-            }}
-          />
+          <div
+            className="bg-image relative h-1/4 bg-cover bg-top bg-no-repeat"
+            style={{ backgroundImage: `url(${imageUrls[0]})` }}
+          >
+            <Button
+              text="To Registration Form"
+              className="absolute right-4 top-4 rounded bg-blue-500 px-4 py-2 text-white hover:bg-blue-600"
+              onClick={() => {
+                setShowEventLandingPage(false);
+              }}
+            />
+          </div>
 
           <div
             className="view ql-editor"
@@ -72,7 +100,11 @@ export default function Template({ params }: { params: PageProps }) {
         </div>
       ) : (
         <>
-          <PreviewForm params={params} />
+          <PreviewForm
+            params={params}
+            imageUrls={imageUrls}
+            setImageUrls={setImageUrls}
+          />
           {previewMode && (
             <Button
               text="To Event Landing Page"
